requirements.txt
python-telegram-bot>=21,<22
requests
